project/requirements.txt
fastapi==0.78.0
fastapi-camelcase==1.0.5
pydantic==1.9.1
pytest==7.1.2
starlette==0.19.1
uvicorn==0.17.6
gunicorn==20.1.0
requests==2.27.1
pytest-env==0.6.2
bcrypt==3.2.2
tortoise-orm==0.19.1
aerich==0.6.3
fastapi-jwt==0.1.8
asyncpg==0.25.0
pytest-asyncio==0.18.3